# docs/openapi/native-v1.yaml
openapi: 3.1.0
...
info:
  title: Odissey Native Client API
  version: 1.0.0
  license:
    name: MIT
    identifier: MIT
  description: |
    Canonical API contract for native Odissey clients. JSON fields use
    camelCase, identifiers are opaque strings, times are UTC ISO 8601, and
    durations and playback positions are integer milliseconds.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh token
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [refreshToken]
              properties:
                refreshToken: { type: string, maxLength: 200 }
      responses:
        '200':
          description: Rotated access and refresh token pair plus active profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Odissey native access token
...
  responses:
...
    Unauthorized:
      description: Missing, invalid, expired, or replayed credentials.
      headers:
        WWW-Authenticate:
          schema: { type: string }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    RateLimited:
      description: Request rate limit exceeded.
      headers:
        Retry-After:
          schema: { type: integer }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
    Problem:
      type: object
      required: [type, title, status, detail, code, requestId]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        code: { type: string }
        requestId:
          type: string
          minLength: 1
          maxLength: 64
          pattern: '^[A-Za-z0-9._:-]+$'
...
    TokenResponse:
      type: object
      required:
        [sessionId, accessToken, refreshToken, tokenType, expiresIn,
         accessExpiresAt, refreshExpiresAt, user, profiles, activeProfile]
      properties:
        sessionId: { type: string }
        accessToken: { type: string, writeOnly: true }
        refreshToken: { type: string, writeOnly: true }
        tokenType: { type: string, const: Bearer }
        expiresIn: { type: integer, minimum: 0 }
        accessExpiresAt: { type: string, format: date-time }
        refreshExpiresAt: { type: string, format: date-time }
        user: { $ref: '#/components/schemas/User' }
        profiles:
          type: array
          items: { $ref: '#/components/schemas/User' }
        activeProfile:
          type: string
          description: ID of the active object in the profiles array.
...
    User:
      type: object
      required: [id, name, email, timezone, isAdmin]
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string, format: email }
        timezone: { type: string }
        isAdmin: { type: boolean }
        active:
          type: boolean
          description: Present for profile-list and authentication payloads.
        preferences: { $ref: '#/components/schemas/Preferences' }
...
    Preferences:
      type: object
      additionalProperties: false
      required: [timezone, autoplay, preferredQuality]
      properties:
        timezone: { type: string }
        autoplay: { type: boolean }
        preferredQuality:
          type: string
          enum: [auto, original, 1080p, 720p]
...